--- a/requirements-dev.lock
+++ b/requirements-dev.lock
@@ -10,6 +10,12 @@
 #   universal: false
 
 -e file:.
+aiohappyeyeballs==2.6.1
+    # via aiohttp
+aiohttp==3.12.11
+    # via turbopuffer
+aiosignal==1.3.2
+    # via aiohttp
 annotated-types==0.6.0
     # via pydantic
 anyio==4.4.0
@@ -17,6 +23,10 @@ anyio==4.4.0
     # via turbopuffer
 argcomplete==3.1.2
     # via nox
+async-timeout==5.0.1
+    # via aiohttp
+attrs==25.3.0
+    # via aiohttp
 certifi==2023.7.22
     # via httpcore
     # via httpx
@@ -34,6 +44,9 @@ execnet==2.1.1
     # via pytest-xdist
 filelock==3.12.4
     # via virtualenv
+frozenlist==1.6.2
+    # via aiohttp
+    # via aiosignal
 h11==0.14.0
     # via httpcore
 httpcore==1.0.2
@@ -44,6 +57,7 @@ httpx==0.28.1
 idna==3.4
     # via anyio
     # via httpx
+    # via yarl
 importlib-metadata==7.0.0
 iniconfig==2.0.0
     # via pytest
@@ -51,6 +65,9 @@ markdown-it-py==3.0.0
     # via rich
 mdurl==0.1.2
     # via markdown-it-py
+multidict==6.4.4
+    # via aiohttp
+    # via yarl
 mypy==1.14.1
 mypy-extensions==1.0.0
     # via mypy
@@ -66,6 +83,9 @@ platformdirs==3.11.0
     # via virtualenv
 pluggy==1.5.0
     # via pytest
+propcache==0.3.1
+    # via aiohttp
+    # via yarl
 pybase64==1.4.1
     # via turbopuffer
 pydantic==2.10.3
@@ -100,12 +120,17 @@ tomli==2.0.2
     # via pytest
 typing-extensions==4.13.2
     # via anyio
+    # via multidict
     # via mypy
     # via pydantic
     # via pydantic-core
     # via pyright
     # via turbopuffer
+urllib3==2.4.0
+    # via turbopuffer
 virtualenv==20.24.5
     # via nox
+yarl==1.20.0
+    # via aiohttp
 zipp==3.17.0
     # via importlib-metadata