--- a/requirements.lock
+++ b/requirements.lock
@@ -48,6 +48,8 @@ idna==3.4
 multidict==6.4.4
     # via aiohttp
     # via yarl
+orjson==3.10.18
+    # via turbopuffer
 propcache==0.3.1
     # via aiohttp
     # via yarl