--- a/requirements.lock
+++ b/requirements.lock
@@ -10,11 +10,21 @@
 #   universal: false
 
 -e file:.
+aiohappyeyeballs==2.6.1
+    # via aiohttp
+aiohttp==3.12.11
+    # via turbopuffer
+aiosignal==1.3.2
+    # via aiohttp
 annotated-types==0.6.0
     # via pydantic
 anyio==4.4.0
     # via httpx
     # via turbopuffer
+async-timeout==5.0.1
+    # via aiohttp
+attrs==25.3.0
+    # via aiohttp
 certifi==2023.7.22
     # via httpcore
     # via httpx
@@ -22,6 +32,9 @@ distro==1.8.0
     # via turbopuffer
 exceptiongroup==1.2.2
     # via anyio
+frozenlist==1.6.2
+    # via aiohttp
+    # via aiosignal
 h11==0.14.0
     # via httpcore
 httpcore==1.0.2
@@ -31,6 +44,13 @@ httpx==0.28.1
 idna==3.4
     # via anyio
     # via httpx
+    # via yarl
+multidict==6.4.4
+    # via aiohttp
+    # via yarl
+propcache==0.3.1
+    # via aiohttp
+    # via yarl
 pybase64==1.4.1
     # via turbopuffer
 pydantic==2.10.3
@@ -42,6 +62,11 @@ sniffio==1.3.0
     # via turbopuffer
 typing-extensions==4.13.2
     # via anyio
+    # via multidict
     # via pydantic
     # via pydantic-core
     # via turbopuffer
+urllib3==2.4.0
+    # via turbopuffer
+yarl==1.20.0
+    # via aiohttp