--- a/src/turbopuffer/_streaming.py
+++ b/src/turbopuffer/_streaming.py
@@ -1,7 +1,6 @@
 # Note: initially copied from https://github.com/florimondmanca/httpx-sse/blob/master/src/httpx_sse/_decoders.py
 from __future__ import annotations
 
-import json
 import inspect
 from types import TracebackType
 from typing import TYPE_CHECKING, Any, Generic, TypeVar, Iterator, AsyncIterator, cast
@@ -9,6 +8,7 @@ from typing_extensions import Self, Protocol, TypeGuard, override, get_origin, r
 
 import httpx
 
+from .lib import json
 from ._utils import extract_type_var_from_base
 
 if TYPE_CHECKING: