--- a/src/turbopuffer/lib/transport.py
+++ b/src/turbopuffer/lib/transport.py
@@ -1,0 +1,1 @@
+MIN_GZIP_SIZE = 1024