… tests/custom/test_compression.py