Also restore request compression for large requests, and client-side performance metrics.

Sync performance:

```
| Benchmark      | bench-baseline | bench-stainless-base   | bench-stainless-opt    | bench-stainless-opt2   | bench-stainless-opt3   |
+================+================+========================+========================+========================+========================+
| upsert         | 1.88 sec       | 896 ms: 2.09x faster   | 1.16 sec: 1.62x faster | 1.12 sec: 1.68x faster | 1.13 sec: 1.66x faster |
+----------------+----------------+------------------------+------------------------+------------------------+------------------------+
| query          | 265 ms         | 277 ms: 1.04x slower   | 340 ms: 1.28x slower   | 252 ms: 1.05x faster   | 244 ms: 1.08x faster   |
+----------------+----------------+------------------------+------------------------+------------------------+------------------------+
| query_scale    | 2.68 sec       | 4.71 sec: 1.76x slower | 4.81 sec: 1.80x slower | 3.37 sec: 1.26x slower | 3.32 sec: 1.24x slower |
+----------------+----------------+------------------------+------------------------+------------------------+------------------------+
| Geometric mean | (ref)          | 1.05x faster           | 1.12x slower           | 1.12x faster           | 1.13x faster           |
+----------------+----------------+------------------------+------------------------+------------------------+------------------------+
```

Base was before this PR. `opt` was the failed experiment to use `urllib3` transports. `opt2` was `urllib3` plus `orjson`. `opt3` is the default httpx transport, plus request compression, plus `orjson`.

There's still a modest slowdown for `query_scale`, but it's tolerable, and the other two are meaningfully faster—`upsert` especially!

Async performance:

```
+----------------+----------------+------------------------+
| Benchmark      | bench-baseline | bench-experiment       |
+================+================+========================+
| query_scale    | 4.34 sec       | 2.57 sec: 1.69x faster |
+----------------+----------------+------------------------+
| Geometric mean | (ref)          | 1.13x faster           |
+----------------+----------------+------------------------+

Benchmark hidden because not significant (2): upsert, query
```

Unlike the sync client, there's no OG async client.

`baseline` is the built in async transport; `experiment` is the aiohttp transport. No significant perf difference in `upsert` or `query`, but the `query_scale` benchmark is a good bit better with the aiohttp transport. Note that I had to disable compression in the aiohttp transport to get a fair comparison (the baseline doesn't do compression), so the absolute speed of the aiohttp transport is not representative of the real world.

Fastest sync vs async performance:

```
+----------------+----------------------+------------------------+
| Benchmark      | bench-stainless-opt3 | bench-experiment       |
+================+======================+========================+
| query          | 244 ms               | 256 ms: 1.05x slower   |
+----------------+----------------------+------------------------+
| query_scale    | 3.32 sec             | 3.00 sec: 1.11x faster |
+----------------+----------------------+------------------------+
| Geometric mean | (ref)                | 1.03x faster           |
+----------------+----------------------+------------------------+

Benchmark hidden because not significant (1): upsert
```

They're basically equivalent in performance.